Moving ready-made mocks under `mocket.mocks` and decorators under `mocket.decorators`.
