+++ mocket/mocks/mockredis.py
